DCLI-1 Modify exec string to print that container information that we need
