--- jest.config.js
@@ -1,5 +1,0 @@
-module.exports = {
-    roots: ['<rootDir>/test'],
-    testRegex: '(/__tests__/.*|(\\.|/)(test|spec))\\.js?$',
-    moduleFileExtensions: ['js', 'json', 'node'],
-}
