+++ package.json
@@ -34,7 +34,10 @@
     "lodash": "^4.17.15"
   },
   "devDependencies": {
+    "chai": "^4.2.0",
+    "mocha": "^8.2.0",
+    "sinon": "^9.2.0",
     "coveralls": "^3.1.0",
-    "jest": "^25.3.0"
+    "jest": "^26.6.1"
   }
 }
