+++ src/commands/startCommand.js
@@ -1,5 +1,5 @@
 const inquirer = require('../inquirer');
-const listActiveContainerNamesPrompt = require('../prompts/stopPrompt');
+const startPrompt = require('../prompts/startPrompt');
 const containerStartExec = require('../execs/containerStartExec');
 const containerStartAllExec = require('../execs/containerStartAllExec');
 
@@ -7,7 +7,7 @@ function startCommand(isAll) {
   if (isAll) {
     containerStartAllExec();
   } else {
-    inquirer.prompt([listActiveContainerNamesPrompt]).then(res => {
+    inquirer.prompt([startPrompt]).then(res => {
       const { type } = res;
       containerStartExec(type);
     });
